requirements.txt
fastapi
uvicorn
celery
redis
Pillow
python-multipart